[SAR-10155] Joda time to Java time
