+++ app/connectors/DataCacheConnector.scala
@@ -28,23 +28,23 @@ import scala.concurrent.{ExecutionContext, Future}
 class DataCacheConnector @Inject()(sessionRepository: SessionRepository, cascadeUpsert: CascadeUpsert)(implicit ec: ExecutionContext) {
 
   def save[A](cacheId: String, key: String, value: A)(implicit fmt: Format[A]): Future[CacheMap] = {
-    sessionRepository().get(cacheId).flatMap { optionalCacheMap =>
+    sessionRepository.get(cacheId).flatMap { optionalCacheMap =>
       val updatedCacheMap = cascadeUpsert(key, value, optionalCacheMap.getOrElse(new CacheMap(cacheId, Map())))
-      sessionRepository().upsert(updatedCacheMap).map { _ => updatedCacheMap }
+      sessionRepository.upsert(updatedCacheMap).map { _ => updatedCacheMap }
     }
   }
 
   def remove(cacheId: String, key: String): Future[Boolean] = {
-    sessionRepository().get(cacheId).flatMap { optionalCacheMap =>
+    sessionRepository.get(cacheId).flatMap { optionalCacheMap =>
       optionalCacheMap.fold(Future(false)) { cacheMap =>
         val newCacheMap = cacheMap copy (data = cacheMap.data - key)
-        sessionRepository().upsert(newCacheMap)
+        sessionRepository.upsert(newCacheMap)
       }
     }
   }
 
   def fetch(cacheId: String): Future[Option[CacheMap]] =
-    sessionRepository().get(cacheId)
+    sessionRepository.get(cacheId)
 
   def getEntry[A](cacheId: String, key: String)(implicit fmt: Format[A]): Future[Option[A]] = {
     fetch(cacheId).map { optionalCacheMap =>
@@ -53,14 +53,14 @@ class DataCacheConnector @Inject()(sessionRepository: SessionRepository, cascade
   }
 
   def addToCollection[A](cacheId: String, collectionKey: String, value: A)(implicit fmt: Format[A]): Future[CacheMap] = {
-    sessionRepository().get(cacheId).flatMap { optionalCacheMap =>
+    sessionRepository.get(cacheId).flatMap { optionalCacheMap =>
       val updatedCacheMap = cascadeUpsert.addRepeatedValue(collectionKey, value, optionalCacheMap.getOrElse(new CacheMap(cacheId, Map())))
-      sessionRepository().upsert(updatedCacheMap).map { _ => updatedCacheMap }
+      sessionRepository.upsert(updatedCacheMap).map { _ => updatedCacheMap }
     }
   }
 
   def removeFromCollection[A](cacheId: String, collectionKey: String, item: A)(implicit fmt: Format[A]): Future[CacheMap] = {
-    sessionRepository().get(cacheId).flatMap { optionalCacheMap =>
+    sessionRepository.get(cacheId).flatMap { optionalCacheMap =>
       optionalCacheMap.fold(throw new Exception(s"Couldn't find document with key $cacheId")) { cacheMap =>
         val newSeq = cacheMap.data(collectionKey).as[Seq[A]].filterNot(x => x == item)
         val newCacheMap = if (newSeq.isEmpty) {
@@ -69,17 +69,17 @@ class DataCacheConnector @Inject()(sessionRepository: SessionRepository, cascade
           cacheMap copy (data = cacheMap.data + (collectionKey -> Json.toJson(newSeq)))
         }
 
-        sessionRepository().upsert(newCacheMap).map { _ => newCacheMap }
+        sessionRepository.upsert(newCacheMap).map { _ => newCacheMap }
       }
     }
   }
 
   def replaceInCollection[A](cacheId: String, collectionKey: String, index: Int, item: A)(implicit fmt: Format[A]): Future[CacheMap] = {
-    sessionRepository().get(cacheId).flatMap { optionalCacheMap =>
+    sessionRepository.get(cacheId).flatMap { optionalCacheMap =>
       optionalCacheMap.fold(throw new Exception(s"Couldn't find document with key $cacheId")) { cacheMap =>
         val newSeq = cacheMap.data(collectionKey).as[Seq[A]].updated(index, item)
         val updatedCacheMap = cacheMap copy (data = cacheMap.data + (collectionKey -> Json.toJson(newSeq)))
-        sessionRepository().upsert(updatedCacheMap).map { _ => updatedCacheMap }
+        sessionRepository.upsert(updatedCacheMap).map { _ => updatedCacheMap }
       }
     }
   }
