+++ app/repositories/SessionRepository.scala
@@ -17,35 +17,21 @@
 package repositories
 
 
-import org.joda.time.{DateTime, DateTimeZone}
+import models.DatedCacheMap
 import org.mongodb.scala.model.Filters.equal
 import org.mongodb.scala.model.Indexes.ascending
 import org.mongodb.scala.model.{FindOneAndReplaceOptions, IndexModel, IndexOptions}
 import play.api.Configuration
-import play.api.libs.json.{Format, JsValue, Json, OFormat}
-import repositories.DatedCacheMap.unapply
 import uk.gov.hmrc.http.cache.client.CacheMap
 import uk.gov.hmrc.mongo.MongoComponent
 import uk.gov.hmrc.mongo.play.json.{Codecs, PlayMongoRepository}
-import uk.gov.hmrc.mongo.play.json.formats.{MongoJavatimeFormats, MongoJodaFormats}
 
 import java.util.concurrent.TimeUnit
-import javax.inject.{Inject, Singleton}
+import javax.inject.Inject
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-case class DatedCacheMap(id: String,
-                         data: Map[String, JsValue],
-                         lastUpdated: DateTime = DateTime.now(DateTimeZone.UTC))
-
-object DatedCacheMap {
-  implicit val dateFormat: Format[DateTime] = MongoJodaFormats.dateTimeFormat //TODO: Temporary, future stories to switch to JavaTime
-  implicit val formats: OFormat[DatedCacheMap] = Json.format[DatedCacheMap]
-
-  def apply(cacheMap: CacheMap): DatedCacheMap = DatedCacheMap(cacheMap.id, cacheMap.data)
-}
-
-class ReactiveMongoRepository(config: Configuration, mongo: MongoComponent)
+class SessionRepository @Inject()(config: Configuration, mongo: MongoComponent)
   extends PlayMongoRepository[DatedCacheMap](
     mongoComponent = mongo,
     collectionName = config.get[String]("appName"),
@@ -68,12 +54,4 @@ class ReactiveMongoRepository(config: Configuration, mongo: MongoComponent)
 
   def get(id: String): Future[Option[CacheMap]] =
     collection.find[CacheMap](equal("id", id)).headOption()
-}
-
-@Singleton
-class SessionRepository @Inject()(config: Configuration, reactiveMongoComponent: MongoComponent) {
-
-  private lazy val sessionRepository: ReactiveMongoRepository = new ReactiveMongoRepository(config, reactiveMongoComponent)
-
-  def apply(): ReactiveMongoRepository = sessionRepository
 }
